+++ HISTORY.rst
@@ -46,6 +46,7 @@ History
     * Added Objectify helper class to navigate JSON attributes (jsbueno).
     * Improve Objectify to allow defaults and get interation (jsbueno).
     * Added attribute traversal capability to Objectify (jsbueno).
+    * Update transition message based on the result of a transition hook, if message key is returned (rudaporto).